refactor: move order services into dedicated modules
